Added histogram.py
This python program allows the user to interact with a dynamic histogram. This histogram is understood as a "beliefs distribution", where each bin's frequency indicates the amount of agents whose belief about a certain topic lies within the bin's domain.
The program allows the user to modify the frequencies of the bins, in order to see how different statistical indicators, including Esteban-Ray's polarization measure, evolve depending on the histogram's configuration.
Lastly, the program also construct a polynomial curve that best fits the histogram configuration, using a spline approximation algorithm.